esp32/requirements.txt
network==0.1
python-dotenv==1.0.0
esptool==4.5
